+++ lib/core/enums/turbo_command_type.dart
@@ -3,13 +3,13 @@ import 'package:ultra_wide_turbo_cli/core/enums/turbo_option.dart';
 import 'package:ultra_wide_turbo_cli/core/enums/turbo_tag_type.dart';
 
 enum TurboCommandType {
-  dartFix,
+  fix,
   clone,
   archive;
 
   List<String> bashCommands() {
     switch (this) {
-      case TurboCommandType.dartFix:
+      case TurboCommandType.fix:
         return [
           '''find . -name "*.dart" \\
     ! -path "./bin/cache/*" \\
@@ -41,15 +41,14 @@ enum TurboCommandType {
     switch (this) {
       case TurboCommandType.clone:
       case TurboCommandType.archive:
+      case TurboCommandType.fix:
         return name;
-      case TurboCommandType.dartFix:
-        return 'dart-fix';
     }
   }
 
   String get help {
     switch (this) {
-      case TurboCommandType.dartFix:
+      case TurboCommandType.fix:
         return 'Format and fix code in lib/ and test/ directories.';
       case TurboCommandType.clone:
         return 'Clone various ultra wide turbo components.';
@@ -60,7 +59,7 @@ enum TurboCommandType {
 
   String get description {
     switch (this) {
-      case TurboCommandType.dartFix:
+      case TurboCommandType.fix:
         return '''
 Formats and fixes Dart code in lib/ and test/ directories.
 - Runs flutter clean & pub get if --clean flag is used
@@ -101,7 +100,7 @@ Example:
 
   List<String> get aliases {
     switch (this) {
-      case TurboCommandType.dartFix:
+      case TurboCommandType.fix:
       case TurboCommandType.clone:
       case TurboCommandType.archive:
         return [];
@@ -110,7 +109,7 @@ Example:
 
   List<TurboFlagType> get flags {
     switch (this) {
-      case TurboCommandType.dartFix:
+      case TurboCommandType.fix:
         return [TurboFlagType.clean];
       case TurboCommandType.clone:
       case TurboCommandType.archive:
@@ -120,7 +119,7 @@ Example:
 
   List<TurboOption> get options {
     switch (this) {
-      case TurboCommandType.dartFix:
+      case TurboCommandType.fix:
         return [];
       case TurboCommandType.clone:
       case TurboCommandType.archive:
@@ -130,7 +129,7 @@ Example:
 
   String? script({required Set<TurboFlagType> activeFlags}) {
     switch (this) {
-      case TurboCommandType.dartFix:
+      case TurboCommandType.fix:
         return [
           if (activeFlags.hasClean)
             ...TurboFlagType.clean.bashCommands(source: this),